addon/requirements.txt
aiohttp==3.9.1
gitpython==3.1.40
pyyaml==6.0.1
python-dotenv==1.0.0
